Added useQueryString hook to new architecture
