+++ login-workflow/src/new-architecture/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './isFirstRender';
 export * from './language-locale-hooks';
+export * from './useQueryString';
